refactor: wrap IconButton in folder
